apps/url-tracking-worker/requirements.txt
newspaper4k==0.9.3
psycopg2-binary==2.9.9
apscheduler==3.10.4
requests==2.31.0
lxml==4.9.3
feedparser==6.0.11
cssselect==1.2.0